so/create loading page with timer
